remove banner but keep setup
